diagram and refactor names and files
