+++ src/components/IsPlaying.tsx
@@ -2,7 +2,7 @@ import { EditorView } from "@codemirror/view";
 import { observer } from "mobx-react-lite";
 import * as React from "react";
 import { ObsidianBridge } from "../obsidian/ObsidianBridge";
-import { AudioStore } from "../player/Player";
+import { AudioStore } from "../player/AudioStore";
 import { IconSpan } from "./IconButton";
 
 export const IsPlaying = observer(
